Move newly public protocols out of `Internal` folder
